\name{plotConditional2D}
\alias{plotConditional2D}
\title{plotConditional2D}
\description{Plot the 2D conditional profile-Density of a sample: calling \code{\link{plotMarginal2D}} with aggregating function max.}
\usage{plotConditional2D(smp, xCol = 2, yCol = 3, intCol = 1, ...)}
\arguments{
  \item{smp}{
}
  \item{xCol}{
}
  \item{yCol}{
}
  \item{intCol}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler}



\seealso{\code{\link{plotMarginal2D}}
\code{\link{twDEMCInt}}}

